Pizza-Drone/app/requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
mlrose==1.3.0
mlrose-hiive==2.2.4
networkx==2.8.5
numpy==1.23.1
orjson==3.7.11
packaging==21.3
pandas==1.4.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
